+++ geofront/masterkey.py
@@ -13,6 +13,11 @@ Master key renewal process:
 
 For more details, see also :class:`TwoPhaseRenewal`.
 
+.. versionchanged:: group
+   ``CloudMasterKeyStore`` is moved from this module to
+   :mod:`geofront.backends.cloud`.
+   See :class:`~.backends.cloud.CloudMasterKeyStore`.
+
 """
 import collections.abc
 import datetime
@@ -21,8 +26,6 @@ import logging
 import os.path
 import threading
 
-from libcloud.storage.base import Container, StorageDriver
-from libcloud.storage.types import ObjectDoesNotExistError
 from paramiko.pkey import PKey
 from paramiko.rsakey import RSAKey
 from paramiko.sftp_client import SFTPClient
@@ -33,8 +36,7 @@ from .keystore import get_key_fingerprint
 from .remote import AuthorizedKeyList, Remote
 from .util import typed
 
-__all__ = ('CloudMasterKeyStore', 'EmptyStoreError',
-           'FileSystemMasterKeyStore', 'MasterKeyStore',
+__all__ = ('EmptyStoreError', 'FileSystemMasterKeyStore', 'MasterKeyStore',
            'PeriodicalRenewal', 'TwoPhaseRenewal',
            'read_private_key_file', 'renew_master_key')
 
@@ -282,95 +284,3 @@ class FileSystemMasterKeyStore(MasterKeyStore):
     @typed
     def save(self, master_key: PKey):
         master_key.write_private_key_file(self.path)
-
-
-class CloudMasterKeyStore(MasterKeyStore):
-    """Store the master key into the cloud object storage e.g. AWS S3_.
-    It supports more than 20 cloud providers through the efforts of Libcloud_.
-    ::
-
-        from geofront.masterkey import CloudMasterKeyStore
-        from libcloud.storage.types import Provider
-        from libcloud.storage.providers import get_driver
-
-        driver_cls = get_driver(Provider.S3)
-        driver = driver_cls('api key', 'api secret key')
-        container = driver.get_container(container_name='my-master-key-bucket')
-        MASTER_KEY_STORE = CloudMasterKeyStore(container)
-
-    :param driver: the libcloud storage driver
-    :type driver: :class:`libcloud.storage.base.StorageDriver`
-    :param container: the block storage container
-    :type container: :class:`libcloud.storage.base.Container`
-    :param object_name: the object name to use
-    :type object_name: :class:`str`
-
-    .. seealso::
-
-       `Object Storage`__ --- Libcloud
-          Storage API allows you to manage cloud object storage and
-          services such as Amazon S3, Rackspace CloudFiles,
-          Google Storage and others.
-
-    .. _S3: http://aws.amazon.com/s3/
-    .. _Libcloud: http://libcloud.apache.org/
-    __ https://libcloud.readthedocs.org/en/latest/storage/
-
-    """
-
-    @typed
-    def __init__(self,
-                 driver: StorageDriver,
-                 container: Container,
-                 object_name: str):
-        self.driver = driver
-        self.container = container
-        self.object_name = object_name
-
-    @typed
-    def load(self) -> PKey:
-        try:
-            obj = self.driver.get_object(self.container.name, self.object_name)
-        except ObjectDoesNotExistError:
-            raise EmptyStoreError()
-        with io.BytesIO() as buffer_:
-            for chunk in self.driver.download_object_as_stream(obj):
-                if isinstance(chunk, str):  # DummyDriver yields str, not bytes
-                    chunk = chunk.encode()
-                buffer_.write(chunk)
-            buffer_.seek(0)
-            with io.TextIOWrapper(buffer_) as tio:
-                return read_private_key_file(tio)
-
-    @typed
-    def save(self, master_key: PKey):
-        with io.StringIO() as buffer_:
-            master_key.write_private_key(buffer_)
-            pem = buffer_.getvalue()
-        self.driver.upload_object_via_stream(
-            self._countable_iterator([pem]),
-            self.container,
-            self.object_name,
-            {'content_type': 'application/x-pem-key'}
-        )
-
-    class _countable_iterator:
-        """libcloud's storage driver takes an iterator as stream,
-        but some drivers e.g. dummy driver try calling :func:`len()`
-        to the iterator.  This adapter workarounds the situation.
-
-        """
-
-        @typed
-        def __init__(self, sequence: collections.abc.Sequence):
-            self.iterator = iter(sequence)
-            self.length = len(sequence)
-
-        def __len__(self):
-            return self.length
-
-        def __iter__(self):
-            return self
-
-        def __next__(self):
-            return next(self.iterator)